fix(news-store): break circular dependency between hydration and index
Extract NewsState and NewsDeps interfaces into types.ts so hydration.ts
imports from types.ts instead of index.ts, breaking the cycle detected
by madge (circular dependency check).
